+++ packages/plugin-iteratively-node/lib/__tests__/smoke.test.ts
@@ -1,9 +1,10 @@
-/* eslint-disable global-require, no-await-in-loop */
+/* eslint-disable no-await-in-loop */
 /* eslint-disable import/no-dynamic-require, import/no-unresolved, import/extensions */
 import fetch from 'node-fetch';
+import Itly from '@itly/sdk';
 import { requireForTestEnv } from '../../../../__tests__/util';
 
-const IterativelyNodePlugin = requireForTestEnv(__dirname);
+const IterativelyPlugin = requireForTestEnv(__dirname);
 
 jest.mock('node-fetch');
 const wait = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
@@ -34,7 +35,7 @@ beforeEach(() => {
   jest.clearAllMocks();
   jest.resetModules();
 
-  itly = require('@itly/sdk').default;
+  itly = new Itly();
 });
 
 afterEach(() => {
@@ -45,7 +46,7 @@ test.each([
   ['production'],
   ['development'],
 ])('should not crash on load (env: %s)', (environment) => {
-  const iterativelyPlugin = new IterativelyNodePlugin(
+  const iterativelyPlugin = new IterativelyPlugin(
     defaultTestApiKey,
     {
       url: defaultTestUrl,
@@ -54,7 +55,7 @@ test.each([
   );
 
   expect(() => {
-    itly.load({
+    itly.load(undefined, {
       environment,
       plugins: [iterativelyPlugin],
     });
@@ -64,7 +65,7 @@ test.each([
 test('should not post if on production', () => {
   const environment = 'production';
 
-  const iterativelyPlugin = new IterativelyNodePlugin(
+  const iterativelyPlugin = new IterativelyPlugin(
     defaultTestApiKey,
     {
       environment,
@@ -73,7 +74,7 @@ test('should not post if on production', () => {
     },
   );
 
-  itly.load({
+  itly.load(undefined, {
     environment,
     plugins: [iterativelyPlugin],
   });
@@ -86,7 +87,7 @@ test('should not post if on production', () => {
 test('should post when flushAt reached', async () => {
   const environment = 'development';
 
-  const iterativelyPlugin = new IterativelyNodePlugin(
+  const iterativelyPlugin = new IterativelyPlugin(
     defaultTestApiKey,
     {
       environment,
@@ -96,7 +97,7 @@ test('should post when flushAt reached', async () => {
     },
   );
 
-  itly.load({
+  itly.load(undefined, {
     environment,
     plugins: [iterativelyPlugin],
   });
@@ -134,7 +135,7 @@ test('should post in flushInterval', async () => {
   const environment = 'development';
   const flushInterval = 10;
 
-  const iterativelyPlugin = new IterativelyNodePlugin(
+  const iterativelyPlugin = new IterativelyPlugin(
     defaultTestApiKey,
     {
       environment,
@@ -144,7 +145,7 @@ test('should post in flushInterval', async () => {
     },
   );
 
-  itly.load({
+  itly.load(undefined, {
     environment,
     plugins: [iterativelyPlugin],
   });
@@ -183,10 +184,59 @@ test('should post in flushInterval', async () => {
   });
 });
 
+test('should post on explicit flush()', async () => {
+  const environment = 'development';
+
+  const iterativelyPlugin = new IterativelyPlugin(
+    defaultTestApiKey,
+    {
+      environment,
+      url: defaultTestUrl,
+    },
+  );
+
+  itly.load(undefined, {
+    environment,
+    plugins: [iterativelyPlugin],
+  });
+
+  const events = [0, 1].map((i) => ({
+    ...defaultTestEvent,
+    properties: {
+      iteration: i,
+    },
+  }));
+
+  for (let i = 0; i < events.length; i += 1) {
+    itly.track(defaultUserId, events[i]);
+  }
+
+  await itly.flush();
+
+  expect(fetch).toHaveBeenCalledTimes(1);
+  expect(fetch).toHaveBeenCalledWith(defaultTestUrl, defaultFetchRequest);
+
+  expect(JSON.parse((fetch as any).mock.calls[0][1].body)).toMatchObject({
+    objects: expect.arrayContaining(
+      events.map((event) => expect.objectContaining({
+        type: 'track',
+        eventId: event.id,
+        eventName: event.name,
+        eventSchemaVersion: event.version,
+        properties: event.properties,
+        valid: true,
+        validation: {
+          details: '',
+        },
+      })),
+    ),
+  });
+});
+
 test('should omit event properties if configured', async () => {
   const environment = 'development';
 
-  const iterativelyPlugin = new IterativelyNodePlugin(
+  const iterativelyPlugin = new IterativelyPlugin(
     defaultTestApiKey,
     {
       environment,
@@ -196,7 +246,7 @@ test('should omit event properties if configured', async () => {
     },
   );
 
-  itly.load({
+  itly.load(undefined, {
     environment,
     plugins: [iterativelyPlugin],
   });
@@ -235,7 +285,7 @@ test('should omit event properties if configured', async () => {
 test('should post track validation error', async () => {
   const environment = 'development';
 
-  const iterativelyPlugin = new IterativelyNodePlugin(
+  const iterativelyPlugin = new IterativelyPlugin(
     defaultTestApiKey,
     {
       environment,
@@ -244,7 +294,7 @@ test('should post track validation error', async () => {
     },
   );
 
-  itly.load({
+  itly.load(undefined, {
     environment,
     plugins: [iterativelyPlugin],
   });
@@ -286,7 +336,7 @@ test('should post track validation error', async () => {
 test('should omit validation error details if configured', async () => {
   const environment = 'development';
 
-  const iterativelyPlugin = new IterativelyNodePlugin(
+  const iterativelyPlugin = new IterativelyPlugin(
     defaultTestApiKey,
     {
       environment,
@@ -296,7 +346,7 @@ test('should omit validation error details if configured', async () => {
     },
   );
 
-  itly.load({
+  itly.load(undefined, {
     environment,
     plugins: [iterativelyPlugin],
   });