+++ packages/plugin-iteratively-node/lib/index.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-unused-vars, class-methods-use-this, no-constant-condition, no-await-in-loop */
 import fetch from 'node-fetch';
 import {
-  Environment, Event, Properties, PluginBase, ValidationResponse,
+  Environment, Event, Properties, Plugin, ValidationResponse,
 } from '@itly/sdk';
 
 export type IterativelyOptions = {
@@ -34,9 +34,10 @@ type TrackModel = {
   };
 };
 
-export default class IterativelyNodePlugin extends PluginBase {
-  static ID: string = 'iteratively';
-
+/**
+ * Iteratively Node Plugin for Iteratively SDK
+ */
+export class IterativelyPlugin extends Plugin {
   private buffer: TrackModel[] = [];
 
   private timer: ReturnType<typeof setTimeout> | null = null;
@@ -52,7 +53,7 @@ export default class IterativelyNodePlugin extends PluginBase {
   };
 
   constructor(private apiKey: string, iterativelyOptions: IterativelyOptions) {
-    super();
+    super('iteratively');
 
     // adjusts config values in accordance with provided environment value
     if (iterativelyOptions.environment === 'production') {
@@ -63,10 +64,7 @@ export default class IterativelyNodePlugin extends PluginBase {
     this.config = { ...this.config, ...iterativelyOptions };
   }
 
-  // overrides PluginBase.id
-  id = () => IterativelyNodePlugin.ID;
-
-  // overrides PluginBase.postIdentify
+  // overrides Plugin.postIdentify
   postIdentify(
     userId: string | undefined,
     properties: Properties | undefined,
@@ -75,7 +73,7 @@ export default class IterativelyNodePlugin extends PluginBase {
     this.push(this.toTrackModel(TrackType.identify, undefined, properties, validationResponses));
   }
 
-  // overrides PluginBase.postGroup
+  // overrides Plugin.postGroup
   postGroup(
     userId: string | undefined,
     groupId: string,
@@ -85,7 +83,7 @@ export default class IterativelyNodePlugin extends PluginBase {
     this.push(this.toTrackModel(TrackType.group, undefined, properties, validationResponses));
   }
 
-  // overrides PluginBase.postPage
+  // overrides Plugin.postPage
   postPage(
     userId: string | undefined,
     category: string | undefined,
@@ -96,7 +94,7 @@ export default class IterativelyNodePlugin extends PluginBase {
     this.push(this.toTrackModel(TrackType.page, undefined, properties, validationResponses));
   }
 
-  // overrides PluginBase.postTrack
+  // overrides Plugin.postTrack
   postTrack(
     userId: string | undefined,
     event: Event,
@@ -143,7 +141,7 @@ export default class IterativelyNodePlugin extends PluginBase {
     return model;
   }
 
-  private async flush() {
+  async flush() {
     if (this.config.disabled) {
       return;
     }
@@ -190,3 +188,5 @@ export default class IterativelyNodePlugin extends PluginBase {
     }
   }
 }
+
+export default IterativelyPlugin;