+++ packages/plugin-iteratively/lib/index.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-unused-vars, class-methods-use-this, no-constant-condition, no-await-in-loop */
 import {
-  Environment, Event, Properties, PluginBase, ValidationResponse,
+  Environment, Event, Properties, Plugin, ValidationResponse,
 } from '@itly/sdk';
 
 export type IterativelyOptions = {
@@ -33,9 +33,10 @@ type TrackModel = {
   };
 };
 
-export default class IterativelyBrowserPlugin extends PluginBase {
-  static ID: string = 'iteratively';
-
+/**
+ * Iteratively Browser Plugin for Iteratively SDK
+ */
+export class IterativelyPlugin extends Plugin {
   private buffer: TrackModel[] = [];
 
   private timer: ReturnType<typeof setTimeout> | null = null;
@@ -51,7 +52,7 @@ export default class IterativelyBrowserPlugin extends PluginBase {
   };
 
   constructor(private apiKey: string, iterativelyOptions: IterativelyOptions) {
-    super();
+    super('iteratively');
 
     // adjusts config values in accordance with provided environment value
     if (iterativelyOptions.environment === 'production') {
@@ -62,10 +63,7 @@ export default class IterativelyBrowserPlugin extends PluginBase {
     this.config = { ...this.config, ...iterativelyOptions };
   }
 
-  // overrides PluginBase.id
-  id = () => IterativelyBrowserPlugin.ID;
-
-  // overrides PluginBase.postIdentify
+  // overrides Plugin.postIdentify
   postIdentify(
     userId: string | undefined,
     properties: Properties | undefined,
@@ -74,7 +72,7 @@ export default class IterativelyBrowserPlugin extends PluginBase {
     this.push(this.toTrackModel(TrackType.identify, undefined, properties, validationResponses));
   }
 
-  // overrides PluginBase.postGroup
+  // overrides Plugin.postGroup
   postGroup(
     userId: string | undefined,
     groupId: string,
@@ -84,7 +82,7 @@ export default class IterativelyBrowserPlugin extends PluginBase {
     this.push(this.toTrackModel(TrackType.group, undefined, properties, validationResponses));
   }
 
-  // overrides PluginBase.postPage
+  // overrides Plugin.postPage
   postPage(
     userId: string | undefined,
     category: string | undefined,
@@ -95,7 +93,7 @@ export default class IterativelyBrowserPlugin extends PluginBase {
     this.push(this.toTrackModel(TrackType.page, undefined, properties, validationResponses));
   }
 
-  // overrides PluginBase.postTrack
+  // overrides Plugin.postTrack
   postTrack(
     userId: string | undefined,
     event: Event,
@@ -142,7 +140,7 @@ export default class IterativelyBrowserPlugin extends PluginBase {
     return model;
   }
 
-  private async flush() {
+  async flush() {
     if (this.config.disabled) {
       return;
     }
@@ -189,3 +187,5 @@ export default class IterativelyBrowserPlugin extends PluginBase {
     }
   }
 }
+
+export default IterativelyPlugin;