+++ packages/plugin-segment/lib/index.ts
@@ -1,15 +1,14 @@
 /* eslint-disable no-unused-vars, class-methods-use-this, import/no-unresolved */
 import {
-  Options, Event, Properties, PluginBase,
+  Options, Event, Properties, Plugin,
 } from '@itly/sdk';
 
 export type SegmentOptions = {};
 
-export default class SegmentBrowserPlugin extends PluginBase {
-  static ID: string = 'segment';
-
-  id = () => SegmentBrowserPlugin.ID;
-
+/**
+ * Segment Browser Plugin for Iteratively SDK
+ */
+export class SegmentPlugin extends Plugin {
   private get segment(): any {
     // eslint-disable-next-line no-restricted-globals
     const s: any = typeof self === 'object' && self.self === self && self;
@@ -20,7 +19,7 @@ export default class SegmentBrowserPlugin extends PluginBase {
     private writeKey: string,
     private options?: SegmentOptions,
   ) {
-    super();
+    super('segment');
   }
 
   load() {
@@ -29,8 +28,8 @@ export default class SegmentBrowserPlugin extends PluginBase {
       // @ts-ignore
       // eslint-disable-next-line
       !function(){var analytics=window.analytics=window.analytics||[];if(!analytics.initialize)if(analytics.invoked)window.console&&console.error&&console.error("Segment snippet included twice.");else{analytics.invoked=!0;analytics.methods=["trackSubmit","trackClick","trackLink","trackForm","pageview","identify","reset","group","track","ready","alias","debug","page","once","off","on","addSourceMiddleware","addIntegrationMiddleware","setAnonymousId","addDestinationMiddleware"];analytics.factory=function(t){return function(){var e=Array.prototype.slice.call(arguments);e.unshift(t);analytics.push(e);return analytics}};for(var t=0;t<analytics.methods.length;t++){var e=analytics.methods[t];analytics[e]=analytics.factory(e)}analytics.load=function(t,e){var n=document.createElement("script");n.type="text/javascript";n.async=!0;n.src="https://cdn.segment.com/analytics.js/v1/"+t+"/analytics.min.js";var a=document.getElementsByTagName("script")[0];a.parentNode.insertBefore(n,a);analytics._loadOptions=e};analytics.SNIPPET_VERSION="4.1.0";}}();
+      this.segment.load(this.writeKey, this.options);
     }
-    this.segment.load(this.writeKey, this.options);
   }
 
   alias(userId: string, previousId: string | undefined) {
@@ -64,3 +63,5 @@ export default class SegmentBrowserPlugin extends PluginBase {
     this.segment.reset();
   }
 }
+
+export default SegmentPlugin;