+++ packages/plugin-snowplow/lib/index.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-unused-vars, class-methods-use-this, import/no-unresolved */
 import {
-  Options, Event, Properties, PluginBase,
+  Options, Event, Properties, Plugin,
 } from '@itly/sdk';
 
 export type SnowplowOptions = {
@@ -8,9 +8,10 @@ export type SnowplowOptions = {
   config?: {};
 };
 
-export default class SnowplowBrowserPlugin extends PluginBase {
-  static ID: string = 'snowplow';
-
+/**
+ * Snowplow Browser Plugin for Iteratively SDK
+ */
+export class SnowplowPlugin extends Plugin {
   private get snowplow(): any {
     // eslint-disable-next-line no-restricted-globals
     const s: any = typeof self === 'object' && self.self === self && self;
@@ -21,10 +22,8 @@ export default class SnowplowBrowserPlugin extends PluginBase {
     readonly vendor: string,
     private options: SnowplowOptions,
   ) {
-    super();
+    super('snowplow');
   }
-
-  id = () => SnowplowBrowserPlugin.ID;
 
   load() {
     if (!this.snowplow) {
@@ -32,8 +31,8 @@ export default class SnowplowBrowserPlugin extends PluginBase {
       // @ts-ignore
       // eslint-disable-next-line
       ;(function(p,l,o,w,i,n,g){if(!p[i]){p.GlobalSnowplowNamespace=p.GlobalSnowplowNamespace||[];p.GlobalSnowplowNamespace.push(i);p[i]=function(){(p[i].q=p[i].q||[]).push(arguments)};p[i].q=p[i].q||[];n=l.createElement(o);g=l.getElementsByTagName(o)[0];n.async=1;n.src=w;g.parentNode.insertBefore(n,g)}}(window,document,"script","//d1fc8wv8zag5ca.cloudfront.net/2.10.2/sp.js","snowplow"));
+      this.snowplow('newTracker', 'itly', this.options.url, this.options.config);
     }
-    this.snowplow('newTracker', 'itly', this.options.url, this.options.config);
   }
 
   identify(userId: string | undefined, properties?: Properties) {
@@ -52,3 +51,5 @@ export default class SnowplowBrowserPlugin extends PluginBase {
     });
   }
 }
+
+export default SnowplowPlugin;