+++ packages/sdk/lib/__tests__/node-smoke.test.ts
@@ -12,12 +12,14 @@ const plugins = [new CustomPlugin()];
 
 export type TestParams = {
   name: string;
+  context: any,
   options: any,
 };
 
 const testParams: TestParams[] = [
   {
     name: 'context = undefined',
+    context: undefined,
     options: {
       environment: 'production',
       plugins,
@@ -25,13 +27,13 @@ const testParams: TestParams[] = [
   },
   {
     name: 'context = 2 properties',
+    context: {
+      requiredString: 'A required string',
+      optionalEnum: 'Value 1',
+    },
     options: {
       environment: 'production',
       plugins,
-      context: {
-        requiredString: 'A required string',
-        optionalEnum: 'Value 1',
-      },
     },
   },
 ];
@@ -46,7 +48,8 @@ describe('should load and track events to a custom destination (no validation)',
   beforeEach(() => {
     jest.resetModules();
 
-    itly = requireForTestEnv(__dirname);
+    const Itly = requireForTestEnv(__dirname);
+    itly = new Itly();
 
     spyConsoleLog = jest.spyOn(console, 'log').mockImplementation();
   });
@@ -57,7 +60,7 @@ describe('should load and track events to a custom destination (no validation)',
   });
 
   test.each(testParams.map((test) => [test.name, test]) as any[])('%s',
-    async (name: string, { options }: TestParams) => {
+    async (name: string, { context, options }: TestParams) => {
       // Try tracking before load, should throw error
       try {
         itly.identify(userId);
@@ -66,11 +69,11 @@ describe('should load and track events to a custom destination (no validation)',
       }
 
       // Load
-      itly.load(options);
+      itly.load(context, options);
 
       // Try load() again, should throw errror
       try {
-        itly.load(options);
+        itly.load(context, options);
       } catch (e) {
         console.log(`Caught expected error. ${e.message}`);
       }
@@ -119,10 +122,6 @@ describe('should load and track events to a custom destination (no validation)',
       // do nothing
       }
 
-      const customOnly = itly.getPlugin('custom');
-      // eslint-disable-next-line no-console
-      console.log('CustomPlugin.id()', customOnly!.id());
-
       expect(spyConsoleLog.mock.calls).toMatchSnapshot();
     });
 
@@ -130,7 +129,7 @@ describe('should load and track events to a custom destination (no validation)',
     const plugin = new CustomPlugin();
     const spyPluginLoad = jest.spyOn(plugin, 'load');
 
-    itly.load({
+    itly.load(undefined, {
       plugins: [plugin],
       disabled: false,
     });
@@ -165,9 +164,8 @@ describe('should load and track events to a custom destination (no validation)',
       }
     }
 
-    itly.load({
+    itly.load(context, {
       environment: 'production',
-      context,
       plugins: [testingPlugin],
     });