+++ packages/sdk/lib/node.ts
@@ -1,12 +1,13 @@
-/* eslint-disable no-unused-vars, class-methods-use-this, import/extensions, import/no-unresolved */
-import {
-  itly as itlyBase,
+/* eslint-disable import/extensions, import/no-unresolved */
+import { Itly } from './base';
+
+export {
+  Itly,
   Options,
   Environment,
   Event,
   Properties,
   Plugin,
-  PluginBase,
   PluginLoadOptions,
   ValidationOptions,
   ValidationResponse,
@@ -14,52 +15,4 @@ import {
   LOGGERS,
 } from './base';
 
-export {
-  Options,
-  Environment,
-  Plugin,
-  PluginBase,
-  PluginLoadOptions,
-  Event,
-  Properties,
-  ValidationOptions,
-  ValidationResponse,
-  Logger,
-  LOGGERS,
-};
-
-// Itly Node SDK is the same as base (for now)
-export class ItlyNode {
-  load = (
-    options: Options,
-  ) => itlyBase.load(options);
-
-  alias = (
-    userId: string, previousId?: string,
-  ) => itlyBase.alias(userId, previousId);
-
-  identify = (
-    userId: string | undefined, identifyProperties?: Properties,
-  ) => itlyBase.identify(userId, identifyProperties);
-
-  group = (
-    userId: string | undefined, groupId: string, groupProperties?: Properties,
-  ) => itlyBase.group(userId, groupId, groupProperties);
-
-  page = (
-    userId: string | undefined, category: string, name: string, pageProperties?: Properties,
-  ) => itlyBase.page(userId, category, name, pageProperties);
-
-  track = (
-    userId: string | undefined,
-    event: Event,
-  ) => itlyBase.track(userId, event);
-
-  reset = () => itlyBase.reset();
-
-  getPlugin = (id: string) => itlyBase.getPlugin(id);
-}
-
-const itly = new ItlyNode();
-export { itly };
-export default itly;
+export default Itly;